add: Archivo para los tipos de Grupo de Etiqueta
