+++ src/server/types/grupoEtiquetas.ts
@@ -1,0 +1,9 @@
+export type GrupoEtiqueta = {
+    id: string;
+    created_at: string;
+    updated_at: string;
+    nombre: string;
+    color: string;
+    esExclusivo: boolean;
+  };
+
